[MS][LITE][Develop] int deconv neon code

+++ mindspore/lite/src/runtime/kernel/arm/nnacl/int8/deconv.h
@@ -29,7 +29,8 @@ extern "C" {
 #endif
 void DeConvPackWeightSum(int8_t *weight, int32_t *weight_sum, int32_t input_zp, int32_t filter_zp, int deep16, int col4,
                          bool suppport_opt);
-void DeConvPackInputSum(const int8_t *src, int32_t *dst, int32_t filter_zp, int row4, int col16, bool suppport_opt);
+void DeConvPackInputSum(const int8_t *src, int32_t *dst, int32_t filter_zp, size_t row4, size_t col16,
+                        bool suppport_opt);
 void DeConvWeightTransInt8(int8_t *src, int8_t *dst, int input_channel, int output_channel, int plane,
                            bool support_optimize_);
